refactor: move restore code with dependencies to broker and dynamic-config to standalone
The old restore code used still in the RestoreApp causes nuance with dependency cycles while creating the new restore procedure. We want to inverse control of the whole restore procedure and this is being a hurdle at the moment.

When the implementation is done we would probably be able to merge these approaches back together.
